modular project structure and minor bug fixes
